requirements/base.txt
flask
Flask-SQLAlchemy
Flask-RESTful
python-dotenv